Move walking functions for goalkeeper to goalkeeper.rs
