+++ crates/nodes/behavior_node/src/walk.rs
@@ -212,114 +212,6 @@ pub fn walk_to_block_position(blackboard: &mut Blackboard) -> Status {
     }
 }
 
-pub fn set_goalkeeper_active_defense_position(blackboard: &mut Blackboard) -> Status {
-    if let (Some(ball), Some(ground_to_field)) = (
-        &blackboard.ball,
-        blackboard.world_state.robot.ground_to_field,
-    ) {
-        let field_dimensions = blackboard.field_dimensions;
-        let parameters = &blackboard.parameters.keeper;
-
-        let own_goal_line_x = -field_dimensions.length / 2.0;
-        let own_goal_center = point!(own_goal_line_x, 0.0);
-        let goal_to_ball = ball.position - own_goal_center;
-        let defense_radius =
-            (field_dimensions.goal_inner_width / 2.0 - parameters.distance_to_goalpost).max(0.0);
-        let unclamped_defense_position_in_field = if goal_to_ball.norm() < f32::EPSILON {
-            point!(own_goal_line_x + defense_radius, 0.0)
-        } else {
-            own_goal_center + goal_to_ball.normalize() * defense_radius
-        };
-        let minimum_defense_x = own_goal_line_x + parameters.distance_to_goalpost;
-        let defense_position_in_field = point!(
-            unclamped_defense_position_in_field
-                .x()
-                .max(minimum_defense_x),
-            unclamped_defense_position_in_field.y()
-        );
-        let defense_position_in_ground = ground_to_field.inverse() * defense_position_in_field;
-
-        if defense_position_in_ground.coords().norm()
-            > parameters.active_defense_maximum_robot_distance
-        {
-            return Status::Failure;
-        }
-
-        blackboard.walk_position = Some(defense_position_in_ground);
-
-        Status::Success
-    } else {
-        Status::Failure
-    }
-}
-
-pub fn walk_to_goalkeeper_default_position(blackboard: &mut Blackboard) -> Status {
-    if let Some(ground_to_field) = blackboard.world_state.robot.ground_to_field {
-        let field_dimensions = blackboard.field_dimensions;
-        let parameters = &blackboard.parameters.keeper;
-        let own_goal_line_x = -field_dimensions.length / 2.0;
-        let default_position_in_field = point!(own_goal_line_x + parameters.x_offset, 0.0);
-        let default_pose_in_field =
-            Pose2::from_parts(default_position_in_field, Orientation2::new(0.0));
-        let default_pose_in_ground = ground_to_field.inverse() * default_pose_in_field;
-        let orientation_mode = if let Some(ball) = &blackboard.world_state.ball {
-            OrientationMode::LookAt {
-                target: ball.ball_in_ground,
-                tolerance: blackboard.parameters.walk_and_stand.orientation_tolerance,
-            }
-        } else if blackboard
-            .ball
-            .as_ref()
-            .is_some_and(|ball| ball.position.x() <= 0.0)
-        {
-            OrientationMode::LookAt {
-                target: ground_to_field.inverse() * point!(own_goal_line_x, 0.0),
-                tolerance: blackboard.parameters.walk_and_stand.orientation_tolerance,
-            }
-        } else {
-            OrientationMode::AlignWithPath
-        };
-
-        walk_to(
-            blackboard,
-            default_pose_in_ground,
-            blackboard.parameters.walk_speed.blocking,
-            orientation_mode,
-            blackboard
-                .parameters
-                .walk_and_stand
-                .normal_distance_to_be_aligned,
-            blackboard.parameters.walk_and_stand.hysteresis,
-        )
-    } else {
-        Status::Failure
-    }
-}
-
-pub fn walk_to_goalkeeper_penalty_position(blackboard: &mut Blackboard) -> Status {
-    if let Some(ground_to_field) = blackboard.world_state.robot.ground_to_field {
-        let own_goal_line_x = -blackboard.field_dimensions.length / 2.0;
-        let penalty_position_in_field = point!(own_goal_line_x, 0.0);
-        let penalty_pose_in_field =
-            Pose2::from_parts(penalty_position_in_field, Orientation2::new(0.0));
-        let penalty_pose_in_ground = ground_to_field.inverse() * penalty_pose_in_field;
-
-        walk_to(
-            blackboard,
-            penalty_pose_in_ground,
-            blackboard.parameters.walk_speed.blocking,
-            OrientationMode::AlignWithPath,
-            blackboard
-                .parameters
-                .walk_and_stand
-                .normal_distance_to_be_aligned,
-            blackboard.parameters.walk_and_stand.hysteresis,
-        )
-    } else {
-        Status::Failure
-    }
-}
-
 pub fn walk_to_kickoff_pose(blackboard: &mut Blackboard) -> Status {
     if let (Some(ground_to_field), player_number) = (
         blackboard.world_state.robot.ground_to_field,
